feat: add platform realtime api
