+++ .devo/provisioning/apps/platform-realtime-api.yml
@@ -1,0 +1,41 @@
+version: "3.9"
+
+services:
+  platform-realtime-api:
+    extends:
+      file: common.yml
+      service: common
+    container_name: platform-realtime-api
+    image: turnly/platform-realtime-api:latest
+    build:
+      context: ${DOCKER_RELATIVE_APPS_DIR}/platform-realtime-api
+      target: ${DOCKER_BUILD_STAGE}
+      dockerfile: Dockerfile
+      args:
+        NODE_ENV: ${NODE_ENV}
+    volumes:
+      - ${DOCKER_RELATIVE_APPS_DIR}/platform-realtime-api:/app
+    depends_on:
+      iam:
+        condition: service_healthy
+    networks:
+      turnly.network.public:
+      turnly.network.internal:
+        aliases:
+          - platform-realtime-api.turnly.local
+    environment:
+      # This service shared env variables that extend from common.yml
+      APP_NAME: platform-realtime-api
+      RABBITMQ_QUEUE: ${PLATFORM_REALTIME_API_RABBITMQ_QUEUE:-turnly.ms.platform_realtime_api.queue}
+    labels:
+      - traefik.enable=true
+      - traefik.docker.network=turnly.network.internal
+
+      # These tags are used to enable authenticated access to the service via OAuth-middleware.
+      - traefik.http.routers.platform-realtime-api.entrypoints=web,websecure
+      - traefik.http.routers.platform-realtime-api.rule=PathPrefix(`/api/rtm/platform`)
+      - traefik.http.routers.platform-realtime-api.service=platform-realtime-api
+      - traefik.http.services.platform-realtime-api.loadbalancer.server.port=${APP_PORT}
+
+      # Attach the OAuth-middleware to this service.
+      - traefik.http.routers.platform-realtime-api.middlewares=oauth-middleware
