+++ apps/platform-realtime-api/.prettierignore
