+++ apps/platform-realtime-api/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@turnly/private-rtm-api",
+  "name": "@turnly/platform-realtime-api",
   "version": "0.0.0",
   "private": true,
   "description": "It's the real-time event bridge that sends data between internal services and applications.",
