+++ apps/platform-realtime-api/src/main.ts
