+++ apps/platform-realtime-api/src/shared/api.ts
