+++ apps/platform-realtime-api/src/shared/config.ts
@@ -9,7 +9,7 @@ import { config } from '@turnly/shared'
 export const serverOptions = Object.freeze({
   name: config.get('app.name'),
   port: config.get('server.port'),
-  path: '/api/rtm/private',
+  path: '/api/rtm/platform',
 })
 
 export enum Channels {
